requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.2
Django==5.0.1
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2023.4
sqlparse==0.4.4
